% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlotMatrix.R
\name{scatterPlotMatrix-shiny}
\alias{scatterPlotMatrix-shiny}
\alias{scatterPlotMatrixOutput}
\alias{renderScatterPlotMatrix}
\title{Shiny bindings for scatterPlotMatrix}
\usage{
scatterPlotMatrixOutput(outputId, width = "100\%", height = "600px")

renderScatterPlotMatrix(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a scatterPlotMatrix}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using scatterPlotMatrix within Shiny
applications and interactive Rmd documents.
}
